fix(web): revert the cursor gate; keep the failures observable (BUG-2508)
Scope call after review: keep this bug narrow, and revert the coordinator change
rather than ship it.

The gate ("advance the cursor only if every consumer applied") was correct at the
service boundary and INERT in production, because no consumer reports failure —
verified at four sites, and confirmed by a live leg that behaved identically on
gated and ungated builds. Shipping it would have read as a fix for the reported
bug while changing nothing, which is worse than the open bug: the next person
cites it as handled. Reverting an inert change loses nothing users ever had.
The design half — consumer contract, the poison-consumer case that would pin the
cursor for everyone against an unbounded /changes window, and the
markSynced/onTabResume inconsistencies — is filed as IDEA-2535.

What ships here, all within "surface the failures":

- FAILURES ARE OBSERVABLE. The old try/catch caught synchronous throws only, and
  two of the five consumers are async with their promise discarded — those
  rejections were not "caught and ignored" but unobserved entirely, surfacing as
  unhandled rejections with nothing tying them to the sync that caused them. A
  rejection handler is now attached to whatever a callback returns, and both arms
  log. Deliberately NOT awaited: consumers keep running concurrently and delivery
  stays synchronous, so observability does not smuggle in an ordering change.
- A `sync_required` ARRIVING MID-SYNC IS DEFERRED, not dropped. The in-flight
  request was issued before that signal, so its window cannot cover it, and
  nothing re-announces the gap.
- Cursor semantics are UNCHANGED, and a test pins that deliberately: if someone
  reinstates the gate, it fails and sends them to IDEA-2535 rather than letting an
  inert contract ship quietly a second time.

Three mutants (drop the async observer, drop the sync log, drop the deferral
flag), each failing only its own test, plus a control that a sync with nothing
pending runs exactly one pass.
